+++ replit.nix
@@ -1,0 +1,8 @@
+{ pkgs }: {
+	deps = [
+		pkgs.nodejs-16_x
+        pkgs.nodePackages.typescript-language-server
+        pkgs.yarn
+        pkgs.replitPackages.jest
+	];
+}
